requirements.txt
python-dotenv
python-multipart
pymongo[srv]
fastapi
uvicorn
pydantic
pytest
httpx
